remove dev deps from production server deployments
