+++ package.json
@@ -37,11 +37,13 @@
     "classnames": "^2.2.5",
     "connect-memcached": "^0.2.0",
     "css-loader": "^0.26.0",
+    "css-modules-require-hook": "^4.0.5",
     "dotenv": "^2.0.0",
     "express": "^4.14.0",
     "express-session": "^1.14.1",
     "express-winston": "^2.0.0",
     "helmet": "^3.1.0",
+    "images-require-hook": "^1.0.3",
     "isomorphic-fetch": "^2.2.1",
     "macaroons.js": "^0.3.6",
     "morgan": "^1.7.0",
@@ -80,7 +82,6 @@
     "chokidar": "^1.6.1",
     "coveralls": "^2.11.14",
     "css-loader": "^0.25.0",
-    "css-modules-require-hook": "^4.0.5",
     "dotenv": "^2.0.0",
     "enzyme": "^2.5.1",
     "eslint": "^3.8.1",
@@ -88,7 +89,6 @@
     "expect": "^1.20.2",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
-    "images-require-hook": "^1.0.3",
     "json-loader": "^0.5.4",
     "mocha": "^3.1.2",
     "nock": "^9.0.2",
